Extract slice functions into dedicated files
