+++ src/gohacks/slice/first_element_or.go
@@ -1,0 +1,9 @@
+package slice
+
+// FirstElementOr provides the first element of the given list or the given alternative if the list is empty.
+func FirstElementOr[C comparable](list []C, alternative C) C { //nolint:ireturn
+	if len(list) > 0 {
+		return list[0]
+	}
+	return alternative
+}
